Mock queries to ModBase.
PCSS queries ModBase via SQL to get AA sequences for UniProt accessions.
This won't work outside of the Sali lab, and will break if any of the
data in ModBase changes, so instead we provide a mock database handle that
returns a snapshot of the database. This should also be faster than
actually querying ModBase.