Leave all job exceptions unhandled
Rather than trying to catch exceptions raised
during postprocessing of the job (explicitly
emailing the admin and giving the user an error
page) leave them unhandled. The framework will
automatically catch these exceptions and fail
the job, so that the admin can fix the underlying
problem without bothering the user.